% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_r_from_phi_and_R.R
\name{A_r_from_phi_and_R}
\alias{A_r_from_phi_and_R}
\title{Calculate A_r from phi and R}
\usage{
A_r_from_phi_and_R(phi_all, R_all)
}
\description{
This function computes A_r matrices for a set of lines given a set of phi and R matrices. Matrices can be given individually or collected into a tensor.
}
