% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_aggregate_and_plot_training.R
\name{accuracy_aggregate_and_plot_training}
\alias{accuracy_aggregate_and_plot_training}
\title{Prediction accuracies aggregator}
\usage{
accuracy_aggregate_and_plot_training(A_all, X_all)
}
\description{
This function takes as input the A_all and X_all tensor objects output by find_DMD_components() and aggregates the iterative and recursive prediction accuracies into a single data frame as well as a longitudinal plot of prediction accuracies across all traits.
}
