% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_timepoints_predictor.R
\name{all_timepoints_predictor}
\alias{all_timepoints_predictor}
\title{Run full time series predictions}
\usage{
all_timepoints_predictor(A_all, X_all)
}
\description{
Function takes two inputs, the A_all and X_all objects output by find_DMD_compoents(), and outputs iterative and recursive predictions for each trait and each time point.
}
