% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DMD_components.R
\name{find_DMD_components}
\alias{find_DMD_components}
\title{Find DMD componets}
\usage{
find_DMD_components(data, r = 25, method = "calculated.A", scale = "min.max")
}
\description{
This function takes as input a data frame in which the first two columns represent two different naming standards for the genotype - "bio_ID" is used and the second column is ignored.
The third column represents the time point, notated here as "DAS".
All subsequent columns are the traits of interest which are then organized into X matrices and used to find A directly or through the Schur-DMD algorithm. These two options are selected by setting method to "calculated.A" or "schur.DMD", respectively.
If "calculated.A" is used the function returns a list of tensors corresponding to the A and X matrices for each line. This corresponds to Algorithm 1 from Paper.
If "Schur.DMD" is used, the function returns a list of tensors corresponding to the components of the Schur-DMD algorithm, this corresponds to Algorithm 2 from Paper.
r indicates the number of singular values and vectors included in the truncated SVD step of Schur-DMD.
scale = "min.max" min-max scales the data for each trait and accession.
}
