% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_tensor.R
\name{table_to_tensor}
\alias{table_to_tensor}
\title{Transform table into tensor}
\usage{
table_to_tensor(table, component = "phi", r = 2, n_timepoints = NULL)
}
\description{
This function takes a table with rows as genotypes and columns as individual matrix entries with column names corresponding to the row and column name in the component matrix separated by "\emph{.}._" and arranges it into a tensor.
}
