% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_unroller.R
\name{tensor_unroller}
\alias{tensor_unroller}
\title{"Unroll" tensors}
\usage{
tensor_unroller(tensor_all)
}
\description{
This function reshapes the tensor objects output by find_DMD_components() into matrices with the genotypes as rows and the individual elements of the component matrices as columns, which can then be used for genomic prediction and heritability analysis.
}
